Refactored Auction Details components
