+++ src/pages/vaults/[vaultid]/auctions/[index].page.tsx
@@ -8,9 +8,10 @@ import {
   LoanVaultLiquidationBatch,
   LoanVaultState
 } from '@defichain/whale-api-client/dist/api/loan'
-import { DesktopAuctionDetails, MobileAuctionDetails } from './_components/AuctionDetails'
 import { AuctionDetailsHeading } from './_components/AuctionDetailsHeading'
 import { EmptySection } from '@components/commons/sections/EmptySection'
+import { DesktopAuctionDetails } from './_components/DesktopAuctionDetails'
+import { MobileAuctionDetails } from './_components/MobileAuctionDetails'
 
 interface ActionsPageProps {
   vault: LoanVaultLiquidated
